Move common parts to a lib directory
Also, add support for fill_local to RowListPayloadGenerator and
EvenRowGenerator
