+++ rowhammer_tester/scripts/playbook/README.md
@@ -42,13 +42,18 @@ The configuration for the row generator.
 
 Number of hammers per row.
 
-*refresh=True|False*
+*refresh=true|false*
 
 Should refresh be enabled?
 
-*verbose=True|False*
+*verbose=true|false*
 
 Should verbose output be generated?
+
+*fill_local=true|false*
+
+Enabling this option permits the RowGenerator to limit the filling and checking of data to a smaller range.  Can improve
+performance.
 
 ### HammerTolerancePayloadGenerator
 
@@ -56,9 +61,13 @@ Payload generator for measuring and characterizing hammer tolerance.  It can pro
 
 The data is generated by doing a series of double-sided hammers against the available group of victim rows.  The double-sided hammers increase in intensity based on *read_count_step*.  Here are the configs that can be specified in *payload_generator_config* for this payload generator:
 
-*verbose=True|False*
+*verbose=true|false*
 
 Should the output be verbose?
+
+*fill_local=true|false*
+
+When enabled, permits shrinking the filled memory area to just the aggressors and the victim.
 
 *row_mapping*
