+++ rowhammer_tester/scripts/playbook/configs/example_hammer_tolerance.cfg
@@ -5,6 +5,7 @@
 	"payload_generator_config" : {
 		"row_mapping" : "TypeARowMapping",
 		"verbose" : true,
+		"fill_local" : true,
 		# If we want 32 victims, we need 34 aggressors
 		"nr_rows" : 34,
 		"nr_chips" : 4,
