+++ rowhammer_tester/scripts/playbook/configs/example_row_list_payload.cfg
@@ -9,6 +9,7 @@
 		"max_iteration" : 10,
 		"verbose" : true,
 		"refresh" : false,
+		"fill_local" : true,
 		"row_generator_config" : {
 			"nr_rows" : 10,
 			"max_row" : 64
