+++ rowhammer_tester/scripts/playbook/row_generators/__init__.py
@@ -13,3 +13,6 @@ class RowGenerator:
 
     def generate_rows(self, iteration):
         raise NotImplementedError("Return a list of rows based on the iteration count")
+
+    def get_memory_range(self, wb, settings):
+        return 0x0, wb.mems.main_ram.size
